moved tests over to relevant file and corrected
